Revert "重构 dbManager.py 为 db_manager.py，从 Class 改为 module"
This reverts commit aec37320224123e70c5df65fdfeaa76f32da4410.
